split spec_builder to smaller modules and remove extra code
